% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bpglm}
\alias{bpglm}
\title{This is main function to estimate the regression coefficients from BIVPOI}
\usage{
bpglm(Y1Y2, X1 = NULL, X2 = NULL, mxit = 150, icob = NULL,
  mtype = 1, ppy = 1, presc = 1e+05, phi = NULL)
}
\arguments{
\item{Y1Y2}{:  Data frame (y1,y2)}

\item{X1}{: Data frame for X1}

\item{X2}{: Data frame for X2}

\item{mxit}{:  Maximum number of iterations}

\item{icob}{:  Vector of initial parameter estimates}

\item{mtype}{: Model type: 1= BIVPOI marginal/marginal; 2= BIVPOI marginal/conditional
3 = ZTBIVPOI marginal/marginal; 4= ZTBIVPOI marginal/conditional
5 = RTBIVPOI marginal/marginal; 6= RTBIVPOI marginal/conditional}

\item{ppy}{: Default value 1 to print results, 0 to suppress print}

\item{presc}{:  Precision of estimates  default value is 100000}

\item{phi}{:  Default "TRUE" Phi's will be estimated; 1 assumes Phis as 1}
}
\value{
Returns a list of estimates and model statistics
}
\description{
This is main function to estimate the regression coefficients from BIVPOI
}
