% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mis_func.R
\name{ChiNA}
\alias{ChiNA}
\title{This is a function to perform Chi-square test.
Takes the input of Log-Likelihood from Null Model and Alternative Model.}
\usage{
ChiNA(NullM, AlterM)
}
\arguments{
\item{NullM}{Reduced model fit results}

\item{AlterM}{Full model fit results}
}
\value{
Provides Chisquare test and p-values
}
\description{
This is a function to perform Chi-square test.
Takes the input of Log-Likelihood from Null Model and Alternative Model.
}
