% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mis_func.R
\name{ChiRF}
\alias{ChiRF}
\title{This is a function to perform Chi-square test.
Takes the input of Log-Likelihood from Full Model and Reduced Model.}
\usage{
ChiRF(Redm, Fullm)
}
\arguments{
\item{Redm}{Reduced model fit results}

\item{Fullm}{Full model fit results}
}
\value{
Provides Chisquare test and p-values
}
\description{
This is a function to perform Chi-square test.
Takes the input of Log-Likelihood from Full Model and Reduced Model.
}
