% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deviance_func.R
\name{deva_func}
\alias{deva_func}
\title{This is a function calculate deviance statistics.}
\usage{
deva_func(y1, y2, Lambdas, phi1, phi2, mtype)
}
\arguments{
\item{y1}{: Outcome variabe y1}

\item{y2}{: Outcome variable y2}

\item{Lambdas}{: Estimated mu1 and mu2 from model}

\item{phi1}{: Estimated phi1}

\item{phi2}{: Estimated phi2}

\item{mtype}{:model type}
}
\value{
Provide deviance
}
\description{
This is a function calculate deviance statistics.
}
