% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dispert1t2_func.R
\name{disper_fun}
\alias{disper_fun}
\title{This is a function to calculate General test for over dispersion
An extension of Hosmer-Lemeshow test for Poisson-Poisson}
\usage{
disper_fun(y1, y2, Lambdas, p, mtype)
}
\arguments{
\item{y1}{: Outcome variabe y1}

\item{y2}{: Outcome variable y2}

\item{Lambdas}{: Lambda estimate from model}

\item{p}{: number of parameters}

\item{mtype}{: model type}
}
\value{
Provides T1 statistic with 2*G d.f.
}
\description{
This is a function to calculate General test for over dispersion
An extension of Hosmer-Lemeshow test for Poisson-Poisson
}
