% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mis_func.R
\name{dpztest}
\alias{dpztest}
\title{This is a function to calculate the z-test for y1 and y2 for dispersion (Impirical). Based on observed y1 and y2.}
\usage{
dpztest(y1, y2, tobs)
}
\arguments{
\item{y1}{: Outcome variabe y1}

\item{y2}{: Outcome variable y2}

\item{tobs}{: Total number of observations n}
}
\value{
Provides  z-value and p-value for test for dispersion
}
\description{
This is a function to calculate the z-test for y1 and y2 for dispersion (Impirical). Based on observed y1 and y2.
}
