% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fir_deri.R
\name{fderiv}
\alias{fderiv}
\title{Function to calculate First derivatives}
\usage{
fderiv(y1, y2, X1, X2, Lambdas, phi1, phi2, tobs, nvar, mtype)
}
\arguments{
\item{y1}{: Outcome variabe y1}

\item{y2}{: Outcome variable y2}

\item{X1}{: Covariates X1}

\item{X2}{: Covariats X2 same as X1}

\item{Lambdas}{: Estimated mu1 and mu2 from model}

\item{phi1}{: Estimated phi1}

\item{phi2}{: Estimated phi2}

\item{tobs}{: Total number of observations n}

\item{nvar}{: Number of covariates ncol(X1)}

\item{mtype}{: model type}
}
\description{
Function to calculate First derivatives
}
