% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Link_func.R
\name{link_func}
\alias{link_func}
\title{Function to calculate mu1 and mu2}
\usage{
link_func(X1, X2, icob1, icob2, mtype)
}
\arguments{
\item{X1}{: Covariates X1}

\item{X2}{: Covariats X2 same as X1}

\item{icob1}{: Beta estimates corresponding to y1}

\item{icob2}{: Beta estimates corresponding to y2}

\item{mtype}{: model type}
}
\value{
return estimated values of Lambdas
}
\description{
Function to calculate mu1 and mu2
}
