% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mis_func.R
\name{PearChiPred}
\alias{PearChiPred}
\title{This is a function to perform Chi-square test for goodness of fit. Based on
the predicted probability.}
\usage{
PearChiPred(y1, y2, Lambdas, nvar, mtype)
}
\arguments{
\item{y1}{: Outcome variabe y1}

\item{y2}{: Outcome variable y2}

\item{Lambdas}{: Estimated mu1 and mu2.}

\item{nvar}{: total number of parameter in bivariate model (including cons.)}

\item{mtype}{: model type}
}
\value{
Provides Chisquare test and p-values
}
\description{
This is a function to perform Chi-square test for goodness of fit. Based on
the predicted probability.
}
