% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phis_func.R
\name{phis_func}
\alias{phis_func}
\title{Function to calculate mu1 and mu2}
\usage{
phis_func(y1, y2, X1, X2, icob1, icob2, tobs, nvar1, nvar2, nvar, mtype)
}
\arguments{
\item{y1}{: Outcome variabe y1}

\item{y2}{: Outcome variable y2}

\item{X1}{: Covariates X1}

\item{X2}{: Covariats X2 same as X1}

\item{icob1}{: Beta estimates corresponding to y1}

\item{icob2}{: Beta estimates corresponding to y2}

\item{tobs}{: Total number of observations n}

\item{nvar1}{: Number of covariates ncol(X1)}

\item{nvar2}{: Number of covariates ncol(X2)}

\item{nvar}{: Number of covariates ncol(X1+X2)}

\item{mtype}{: model type}
}
\value{
return estimated values of phi1 and phi2
}
\description{
Function to calculate mu1 and mu2
}
