% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chk_fun.R
\name{poichkfun}
\alias{poichkfun}
\title{Function to check function arguments}
\usage{
poichkfun(dmat, mxit = 150, mtype)
}
\arguments{
\item{dmat}{:  Data frame (y1,y2,X1,X2)}

\item{mxit}{:  Maximum number of iterations}

\item{mtype}{: Model type: 1= Double Poisson with proposed NC; 2= Efron's Double Poisson Model-1; 3= Efron's Double Poisson Model-2}
}
\value{
returns the data matrix after casewise deletion incase of missing values
#@importFrom("stats", "complete.cases", "na.omit")
}
\description{
Function to check function arguments
}
