% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sec_deri.R
\name{sderiv}
\alias{sderiv}
\title{Function to calculate Second derivatives}
\usage{
sderiv(y1, y2, X1, X2, Lambdas, phi1, phi2, tobs, nvar1, nvar2, nvar,
  mtype)
}
\arguments{
\item{y1}{: Outcome variabe y1}

\item{y2}{: Outcome variable y2}

\item{X1}{: Covariates X1}

\item{X2}{: Covariats X2 same as X1}

\item{Lambdas}{: Estimated mu1 and mu2 from model}

\item{phi1}{: Estimated phi1}

\item{phi2}{: Estimated phi2}

\item{tobs}{: Total number of observations n}

\item{nvar1}{: Number of covariates ncol(X1)}

\item{nvar2}{: Number of covariates ncol(X2)}

\item{nvar}{: Number of covariates ncol(X1+X2)}

\item{mtype}{: model type}
}
\value{
ith second derivative matrix
}
\description{
Function to calculate Second derivatives
}
