% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1t2.R
\name{t1t2}
\alias{t1t2}
\title{This is a function to calculate General test for over dispersion
An extension of Hosmer-Lemeshow test for Poisson-Poisson}
\usage{
t1t2(y1, y2, Lambdas, phi1, phi2, mtype)
}
\arguments{
\item{y1}{: Outcome variabe y1}

\item{y2}{: Outcome variable y2}

\item{Lambdas}{: Lambda estimate from model}

\item{phi1}{: Estimated Phi1}

\item{phi2}{: Estimated Phi2}

\item{mtype}{: model type}
}
\value{
Provides T1 statistic with 2*G d.f.
}
\description{
This is a function to calculate General test for over dispersion
An extension of Hosmer-Lemeshow test for Poisson-Poisson
}
