% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mis_func.R
\name{vountest}
\alias{vountest}
\title{This is a function to perform Chi-square test.
Input saved results from Model 1 and Model 2.}
\usage{
vountest(Mod1, Mod2)
}
\arguments{
\item{Mod1}{Model 1 fit results}

\item{Mod2}{Model 2 fit results}
}
\value{
ProvidesVoung Z-value and p-value
}
\description{
This is a function to perform Chi-square test.
Input saved results from Model 1 and Model 2.
}
